Feature: clean code
