+++ internal/service/runtime.go
@@ -2,10 +2,26 @@ package service
 
 import (
 	"github.com/itering/scale.go/types"
+	"github.com/itering/subscan/model"
 	"github.com/itering/substrate-api-rpc/metadata"
 	"github.com/itering/substrate-api-rpc/rpc"
 	"strings"
 )
+
+func (s *Service) SubstrateRuntimeList() []model.RuntimeVersion {
+	return s.dao.RuntimeVersionList()
+}
+
+func (s *Service) SubstrateRuntimeInfo(spec int) *types.MetadataStruct {
+	if metadataInstant, ok := metadata.RuntimeMetadata[spec]; ok {
+		return metadataInstant
+	}
+	runtime := metadata.Process(s.dao.RuntimeVersionRaw(spec))
+	if runtime == nil {
+		return metadata.Latest(nil)
+	}
+	return runtime
+}
 
 func (s *Service) regRuntimeVersion(name string, spec int) error {
 	if affected := s.dao.CreateRuntimeVersion(name, spec); affected > 0 {
