--- internal/service/scan/runtime.go
@@ -1,22 +1,0 @@
-package scan
-
-import (
-	"github.com/itering/scale.go/types"
-	"github.com/itering/subscan/model"
-	"github.com/itering/substrate-api-rpc/metadata"
-)
-
-func (s *Service) SubstrateRuntimeList() []model.RuntimeVersion {
-	return s.dao.RuntimeVersionList()
-}
-
-func (s *Service) SubstrateRuntimeInfo(spec int) *types.MetadataStruct {
-	if metadataInstant, ok := metadata.RuntimeMetadata[spec]; ok {
-		return metadataInstant
-	}
-	runtime := metadata.Process(s.dao.RuntimeVersionRaw(spec))
-	if runtime == nil {
-		return metadata.Latest(nil)
-	}
-	return runtime
-}
